got createtripform component added and tested it and it goes to database. Also added in the route to take you there
