+++ client/package.json
@@ -40,6 +40,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-    
   }
 }
