+++ client/src/App.js
@@ -1,14 +1,10 @@
 import React from 'react';
-
-import { ApolloClient, InMemoryCache, ApolloProvider, HttpLink } from '@apollo/client';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-
-
 import './App.css';
-import Home from './components/Home'
+import Home from './components/Home';
 import SignUp from './components/SignUp';
 import Login from './components/Login';
-import CreateTripForm from './components/YourTrip/YourTripDetails/CreateTripForm';
+import CreateTripForm from './components/LoggedIn/YourTripDetails/CreateTripForm';
 
 import { ApolloClient, InMemoryCache, ApolloProvider, HttpLink } from '@apollo/client';
 import Navbar from './components/Navbar';
@@ -36,9 +32,12 @@ function App() {
               path="/sign-up" 
               element={<SignUp />}
           />
+             <Route
+              path="/yourtrip" 
+              element={<CreateTripForm />}
+          />
         </Routes>
       </Router>
-      <CreateTripForm />
     </ApolloProvider>
   );
 }
